re_framework: extract filesystem impl into persistence module
Move persist_state/load_state/delete_state (+ private entity_path) out of
handle.rs into persistence.rs, the sole owner of filesystem access. handle.rs
no longer touches the fs or anyhow::Context.

delete_state now returns Result (symmetric with persist_state; NotFound is
Ok), so persistence stays a leaf module with no dependency on handle's
logger — delete logs at the call site. Also rename post-transition commit
helper persist_and_fire -> post_transition.

This is the store API seam: swapping in a real DB later means replacing the
three persistence fns, handle.rs unchanged.
